backend/requirements.txt
fastapi
uvicorn
textblob
scikit-learn
python-docx
PyPDF2
python-multipart
